% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind}
\alias{bind}
\title{Combining \pkg{sp} class objects}
\usage{
\S4method{c}{SpatialPoints}(x, ..., ignore.mcols = FALSE, recursive = FALSE)
}
\arguments{
\item{x, ...}{Multiple \linkS4class{SpatialPoints} objects of the same dimensionality.}

\item{ignore.mcols, recursive}{Further arguments passed to \code{\link{c,Vector-method}}.}
}
\value{
An object of the same class as \code{x} containing coordinates from all objects in \code{...}.
}
\description{
Utilities to combine instances of \pkg{sp} classes.
Perhaps these would have a better home in \pkg{sp} itself,
but until then, here we are.
}
\examples{
coords <- matrix(runif(30), nrow=10)
X <- sp::SpatialPoints(coords)

coords2 <- matrix(runif(15), nrow=5)
Y <- sp::SpatialPoints(coords2)

c(X, Y, X, Y) 

}
\author{
Aaron Lun
}
