% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMoran.R
\name{computeMoran}
\alias{computeMoran}
\title{Compute Moran's I statistic}
\usage{
computeMoran(
  x,
  coordinates,
  prop = 0.05,
  radius = NULL,
  kernel = c("uniform", "triangular", "biweight", "tricube"),
  iterations = 0,
  BPPARAM = SerialParam()
)
}
\description{
Compute Moran's I statistic to measure the strength of spatial associations for each gene.
}
\examples{
coordinates <- cbind(rep(1:10, each=10), rep(1:10, 10))

# Completely random:
x <- matrix(rnorm(1000), ncol=100)
stats <- computeMoran(x, coordinates, radius=3)
stats

# Strongly locally correlated:
hotspots <- coordinates[c(10, 25, 60, 85),]
for (i in seq_len(nrow(hotspots))) {
    in.range <- sqrt(colSums((t(coordinates) - hotspots[i,])^2))
    x[1,in.range <= 1] <- x[1,in.range <= 1] + 10
}
stats2 <- computeMoran(x, coordinates, radius=3)
stats2

}
