% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandPoints.R
\name{expandPoints}
\alias{expandPoints}
\title{Expand points to near-circles}
\usage{
expandPoints(x, radius, resolution = 20)
}
\arguments{
\item{x}{A two-dimensional \linkS4class{SpatialPoints} object, where each point is used as the center of a circle.}

\item{radius}{A numeric vector of length equal to \code{x}, specifying the radius for each circle.
Shorter vectors are recycled as necessary.}

\item{resolution}{Integer scalar indicating the number of vertices of the circumscribed polygon.}
}
\value{
A \linkS4class{SpatialPolygons} object containing the circumscribed polygons.
}
\description{
Expand points to circumscribed polygons to approximate circles of a given radius.
}
\details{
On occasion, it is helpful to convert points into circles around those points,
e.g., to check for overlaps of a minimum distance or overlap.
This function uses circumscribed polygons to approximate circles to the desired resolution,
given that \pkg{sp} does not (seem to?) have a natural way of representing pure circles.
}
\examples{
coords <- matrix(runif(30), nrow=10)
X <- sp::SpatialPoints(coords)
Y <- expandPoints(X, runif(10))
plot(Y)

}
\seealso{
\code{\link{makeSpatialPolygons}}, for more general construction of \linkS4class{SpatialPolygons} objects.
}
\author{
Aaron Lun
}
