% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpatialPolygons.R
\name{makeSpatialPolygons}
\alias{makeSpatialPolygons}
\title{Construct SpatialPolygons}
\usage{
makeSpatialPolygons(..., hole = NULL, proj4string = NULL)
}
\arguments{
\item{...}{One or more 2-dimensional numeric matrices of coordinates, 
where each matrix represents a polygon to be treated as a separate entry in the \linkS4class{SpatialPolygons}.}

\item{hole}{See \code{\link{Polygon}}; if \code{NULL}, that function's default value is used.}

\item{proj4string}{See \code{\link{SpatialPolygons}}; if \code{NULL}, that function's default value is used.}
}
\value{
A \linkS4class{SpatialPolygons} object containing the specified polygons.
}
\description{
A convenience function to construct \linkS4class{SpatialPolygons} from a list of coordinate matrices.
}
\details{
For simple polygons, this utility avoids the hassle of creating a \linkS4class{Polygon},
and then creating a \linkS4class{Polygons}, and then \linkS4class{SpatialPolygons}.
However, more complex cases (involving several polygons per entry of the \linkS4class{SpatialPolygons} object)
will have to be constructed manually.
}
\examples{
spm <- makeSpatialPolygons(
    rbind(c(0,0), c(0.5, 0.2), c(0.2, 0.6), c(0.9, 0.8)),
    rbind(c(1,1), c(0.3, 0.7), c(0.1, 0.2), c(0.8, 0.3))
)
spm

}
\author{
Aaron Lun
}
