% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAsCell.R
\name{showAsCell}
\alias{showAsCell}
\alias{showAsCell,SpatialPoints-method}
\alias{showAsCell,SpatialPolygons-method}
\title{Showing \pkg{sp} classes}
\usage{
\S4method{showAsCell}{SpatialPoints}(object)

\S4method{showAsCell}{SpatialPolygons}(object)
}
\arguments{
\item{object}{An instance of an \pkg{sp} class with vector-like characteristics,
e.g., a \linkS4class{SpatialPointsDataFrame}.}
}
\value{
A character vector containing a representation of each element of \code{object}.
}
\description{
Implements dedicated \code{\link{showAsCell}} methods for the various \pkg{sp} classes
so that they play nice with Bioconductor objects,
particularly when being stored as \code{\link{DFRame}} columns.
}
\examples{
coords <- matrix(runif(30), 10)
X <- sp::SpatialPoints(coords)
showAsCell(X)

df <- DataFrame(location=I(X), other=LETTERS[1:10])
df

# Works for spatial polygons as well, based on the polygon centroid.
spm <- makeSpatialPolygons(matrix(runif(10), ncol=2), matrix(runif(10), ncol=2))
DataFrame(poly=I(spm))

}
\author{
Aaron Lun
}
