% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatula-pkg.R
\name{spatula-pkg}
\alias{spatula-pkg}
\title{The \pkg{spatula} package}
\description{
Provides bindings of \linkS4class{Spatial} classes from the \pkg{sp} package
so that they work nicely with Bioconductor containers and conventions.
This aims to facilitate analysis of biological spatial data (e.g., spatial transcriptomics,
mass cytometry, various types of image analysis) within the Bioconductor framework.
Additional quality-of-life utilities are provided to make it easier to interact with such objects,
such as quick creation of various types of \linkS4class{SpatialPolygons}.
}
\author{
Aaron Lun, Kevin Rue-Albrecht
}
