% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\docType{methods}
\name{spatula-bind}
\alias{spatula-bind}
\alias{bindROWS,SpatialPoints-method}
\alias{bindROWS,SpatialPolygons-method}
\alias{extractROWS,SpatialPolygons-method}
\alias{extractROWS,SpatialPolygons,ANY-method}
\alias{showAsCell,SpatialPoints-method}
\alias{showAsCell,SpatialPolygons-method}
\alias{bindROWS,.PointsVector-method}
\alias{bindROWS,.PolygonsVector-method}
\title{S4Vectors bindings for Spatial objects}
\description{
Implements various internal methods for \linkS4class{SpatialPoints} and \linkS4class{SpatialPolygons} objects,
so that they are compatible with Bioconductor containers - see the \pkg{S4Vectors} package for more details.
}
\examples{
# Storing Spatial classes in containers:
coords <- matrix(runif(30), nrow=10)
X <- DataFrame(location=I(sp::SpatialPoints(coords)))
X

coords2 <- matrix(runif(15), nrow=5)
Y <- DataFrame(location=I(sp::SpatialPoints(coords2)))
Y

# Subsetting and such just works:
rbind(X, Y)
X[1:10,,drop=FALSE]

# Works for spatial polygons as well, based on the polygon centroid.
spm <- makeSpatialPolygons(matrix(runif(10), ncol=2), matrix(runif(10), ncol=2))
pX <- DataFrame(poly=I(spm))
pX

spm2 <- makeSpatialPolygons(matrix(runif(10), ncol=2), matrix(runif(10), ncol=2))
pY <- DataFrame(poly=I(spm2))
pY

combined <- rbind(pX, pY)
combined
combined[c(1,1,2,3,3),,drop=FALSE]

}
\author{
Aaron Lun
}
